created tests project, removed redundant files
