+++ ObjectPrinting/PropertyPrintingConfig.cs
@@ -1,0 +1,30 @@
+﻿using System.Globalization;
+
+namespace ObjectPrinting;
+
+public class PropertyPrintingConfig<TOwner, TPropType> : IPropertyPrintingConfig<TOwner, TPropType>
+{
+    private readonly PrintingConfig<TOwner> printingConfig;
+
+    public PropertyPrintingConfig(PrintingConfig<TOwner> printingConfig)
+    {
+        this.printingConfig = printingConfig;
+    }
+
+    PrintingConfig<TOwner> IPropertyPrintingConfig<TOwner, TPropType>.ParentConfig => printingConfig;
+
+    public PrintingConfig<TOwner> Using(Func<TPropType, string> print)
+    {
+        return printingConfig;
+    }
+
+    public PrintingConfig<TOwner> Using(CultureInfo culture)
+    {
+        return printingConfig;
+    }
+}
+
+public interface IPropertyPrintingConfig<TOwner, TPropType>
+{
+    PrintingConfig<TOwner> ParentConfig { get; }
+}
