--- ObjectPrinting/Solved/ObjectExtensions.cs
@@ -1,10 +1,0 @@
-namespace ObjectPrinting.Solved
-{
-    public static class ObjectExtensions
-    {
-        public static string PrintToString<T>(this T obj)
-        {
-            return ObjectPrinter.For<T>().PrintToString(obj);
-        }
-    }
-}
