+++ ObjectPrintingTests/ObjectPrinterAcceptanceTests.cs
@@ -1,0 +1,39 @@
+﻿using System.Globalization;
+using NUnit.Framework;
+using ObjectPrinting;
+
+namespace ObjectPrintingTests;
+
+[TestFixture]
+public class ObjectPrinterAcceptanceTests
+{
+    [Test]
+    public void Demo()
+    {
+        var person = new Person { Name = "Alex", Age = 19 };
+
+        var printer = ObjectPrinter.For<Person>()
+            //1. Исключить из сериализации свойства определенного типа
+            .Excluding<Guid>()
+            //2. Указать альтернативный способ сериализации для определенного типа
+            .Printing<int>().Using(i => i.ToString("X"))
+            //3. Для всех типов, имеющих культуру, указать культуру
+            .Printing<double>().Using(CultureInfo.InvariantCulture)
+            //4. Настроить сериализацию конкретного свойства
+            //5. Настроить обрезание строковых свойств (метод должен быть виден только для строковых свойств)
+            .Printing(p => p.Name).TrimmedToLength(10)
+            //6. Исключить из сериализации конкретного свойства
+            .Excluding(p => p.Age);
+
+        var s1 = printer.PrintToString(person);
+
+        //7. Синтаксический сахар в виде метода расширения, сериализующего по-умолчанию
+        var s2 = person.PrintToString();
+
+        //8. ...с конфигурированием
+        var s3 = person.PrintToString(s => s.Excluding(p => p.Age));
+        Console.WriteLine(s1);
+        Console.WriteLine(s2);
+        Console.WriteLine(s3);
+    }
+}
