+++ ObjectPrintingTests/Person.cs
@@ -1,0 +1,9 @@
+﻿namespace ObjectPrintingTests;
+
+public class Person
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+    public double Height { get; set; }
+    public int Age { get; set; }
+}
